Todo next: PDF Generation

+++ resources/views/users/users_list.blade.php
@@ -74,9 +74,15 @@
 						                <a href="{{ route('users.edit',$user->id) }}" role="button" class="btn btn-default">
 						                <i class="fa fa-pencil"></i> 
 						                </a>
-						                {!! Form::model($user, ['method'=>'DELETE','action' => ['user\UserController@destroy',$user->id] , 'class' => 'form-horizontal form-label-left form-wrapper']) !!}
-						                    <button type="submit" class="btn btn-default" onclick="return confirm('Are you sure you want to delete this item?');"><i class="fa fa-trash"></i> </button>
-						               	{!! Form::close() !!}
+						                @if(Auth::user()->id != $user->id)
+						                	@if($user->is_active)
+						                		{!! Form::model($user, ['method'=>'DELETE','action' => ['user\UserController@destroy',$user->id] , 'class' => 'form-horizontal form-label-left form-wrapper']) !!}
+							                    	<button type="submit" class="btn btn-default" onclick="return confirm('Are you sure you want to lock this user?');"><i class="fa fa-lock"></i> </button>
+							               		{!! Form::close() !!}
+							               	@else
+							               		<a href="{{route('users.resetpassword',$user->id)}}" class="btn btn-default"><i class="fa fa-unlock"></i></a>
+						                	@endif
+						               	@endif
 						            </td>
 				          		</tr>
 				          	@endforeach
